+++ src/core/option/list-option.component.ts
@@ -10,17 +10,22 @@ import {
     HostListener,
     Optional
 } from '@angular/core';
-import { FocusableOption, FocusOrigin } from '@angular/cdk/a11y';
-import { SelectionModel } from '@angular/cdk/collections';
+import { FocusableOption, FocusOrigin, Highlightable } from '@angular/cdk/a11y';
+// import { SelectionModel } from '@angular/cdk/collections';
 import { inputValueToBoolean } from '../../util/helpers';
 
 let _uniqueIdCounter = 0;
 
 export interface IThyOptionParentComponent {
     multiple?: boolean;
-    selectionModel: SelectionModel<ThyListOptionComponent>;
+    // selectionModel: SelectionModel<ThyListOptionComponent>;
+    // 选择，取消选择 option
     toggleOption(option: ThyListOptionComponent, event?: Event): void;
-    setFocusedOption(option: ThyListOptionComponent, event?: Event): void;
+    // 设置当前选项为激活状态，即 hover 状态
+    setActiveOption(option: ThyListOptionComponent, event?: Event): void;
+    // 滚动到当前的选项
+    scrollIntoView(option: ThyListOptionComponent): void;
+    isSelected(option: ThyListOptionComponent): boolean;
 }
 
 /**
@@ -34,7 +39,7 @@ export const THY_OPTION_PARENT_COMPONENT =
     selector: 'thy-list-option,[thy-list-option]',
     templateUrl: './list-option.component.html'
 })
-export class ThyListOptionComponent implements FocusableOption {
+export class ThyListOptionComponent implements Highlightable {
 
     @HostBinding(`class.thy-list-item`) _isListOption = true;
 
@@ -46,13 +51,18 @@ export class ThyListOptionComponent implements FocusableOption {
 
     @Input() thyValue: any;
 
+    // @Input() thyValueKey: string;
+
     disabled?: boolean;
 
     /** Whether the option is selected. */
-    @HostBinding(`class.active`) selected = false;
+    @HostBinding(`class.active`)
+    get selected() {
+        return this.parentSelectionList.isSelected(this);
+    }
 
     constructor(
-        private element: ElementRef<HTMLElement>,
+        public element: ElementRef<HTMLElement>,
         private changeDetector: ChangeDetectorRef,
         /** @docs-private */
         @Optional() @Inject(THY_OPTION_PARENT_COMPONENT) public parentSelectionList: IThyOptionParentComponent
@@ -63,35 +73,28 @@ export class ThyListOptionComponent implements FocusableOption {
     @HostListener('click', ['$event'])
     onClick(event: Event) {
         this.parentSelectionList.toggleOption(this, event);
+        this.parentSelectionList.setActiveOption(this);
     }
 
-    @HostListener('focus', ['$event'])
-    onFocus(event: Event) {
-        this.parentSelectionList.setFocusedOption(this, event);
-    }
-
-    toggle(): void {
-        this.selected = !this.selected;
-    }
+    // @HostListener('focus', ['$event'])
+    // onFocus(event: Event) {
+    //     this.parentSelectionList.setFocusedOption(this, event);
+    // }
 
     /** Allows for programmatic focusing of the option. */
-    focus(origin?: FocusOrigin): void {
-        this.element.nativeElement.focus();
+    // focus(origin?: FocusOrigin): void {
+    //     this.element.nativeElement.focus();
+    // }
+
+    setActiveStyles(): void {
+        this.element.nativeElement.classList.add('hover');
+        this.parentSelectionList.scrollIntoView(this);
     }
 
-    setSelected(selected: boolean): boolean {
-        if (selected === this.selected) {
-            return false;
-        }
-        this.selected = selected;
-        if (selected) {
-            this.parentSelectionList.selectionModel.select(this);
-        } else {
-            this.parentSelectionList.selectionModel.deselect(this);
-        }
-        this.changeDetector.markForCheck();
-        return true;
+    setInactiveStyles(): void {
+        this.element.nativeElement.classList.remove('hover');
     }
+
     /**
      * Returns the list item's text label. Implemented as a part of the FocusKeyManager.
      * @docs-private