+++ src/list/selection/selection-list.ts
@@ -16,20 +16,20 @@ import {
     NgZone,
     forwardRef
 } from '@angular/core';
-import { FocusKeyManager } from '@angular/cdk/a11y';
+import { FocusKeyManager, ActiveDescendantKeyManager } from '@angular/cdk/a11y';
 import { SelectionModel } from '@angular/cdk/collections';
 import {
     ThyListOptionComponent,
     THY_OPTION_PARENT_COMPONENT,
     IThyOptionParentComponent
 } from '../../core/option';
-import { keycodes, helpers } from '../../util';
+import { keycodes, helpers, dom } from '../../util';
 import { inputValueToBoolean } from '../../util/helpers';
 import { Subscription, throwError } from 'rxjs';
 import { DOCUMENT } from '@angular/platform-browser';
 import { ThySelectionListChange } from './selection.interface';
 import { NG_VALUE_ACCESSOR, ControlValueAccessor } from '@angular/forms';
-
+import { ScrollToService } from '../../core';
 
 @Component({
     selector: 'thy-selection-list,[thy-selection-list]',
@@ -49,7 +49,7 @@ import { NG_VALUE_ACCESSOR, ControlValueAccessor } from '@angular/forms';
 })
 export class ThySelectionListComponent implements OnInit, OnDestroy, AfterContentInit, IThyOptionParentComponent, ControlValueAccessor {
 
-    private _keyManager: FocusKeyManager<ThyListOptionComponent>;
+    private _keyManager: ActiveDescendantKeyManager<ThyListOptionComponent>;
 
     private _selectionChangesUnsubscribe$ = Subscription.EMPTY;
 
@@ -58,7 +58,7 @@ export class ThySelectionListComponent implements OnInit, OnDestroy, AfterConten
     private _tempValues: any[];
 
     /** The currently selected options. */
-    selectionModel: SelectionModel<ThyListOptionComponent>;
+    selectionModel: SelectionModel<any>;
 
     disabled: boolean;
 
@@ -77,15 +77,18 @@ export class ThySelectionListComponent implements OnInit, OnDestroy, AfterConten
         this.multiple = inputValueToBoolean(value);
         if (previousValue !== this.multiple) {
             this._instanceSelectionModel();
-            this._subscribeSelectionModelChanges();
         }
     }
 
     @Input() thyBindKeyEventContainer: HTMLElement | ElementRef | string;
 
+    @Input() thyScrollContainer: HTMLElement | ElementRef | string;
+
     @Input() thyBeforeKeydown: (event?: KeyboardEvent) => boolean;
 
-    @Input() thyCompareWith: (o1: any, o2: any) => boolean;
+    @Input() thyUniqueKey: string;
+
+    @Input() thyCompareWith: ((o1: any, o2: any) => boolean);
 
     /** Emits a change event whenever the selected state of an option changes. */
     @Output() readonly thySelectionChange: EventEmitter<ThySelectionListChange> =
@@ -98,9 +101,30 @@ export class ThySelectionListComponent implements OnInit, OnDestroy, AfterConten
     private _emitChangeEvent(option: ThyListOptionComponent, event: Event) {
         this.thySelectionChange.emit({
             source: this,
+            value: option.thyValue,
             option: option,
-            event: event
+            event: event,
+            selected: this.isSelected(option)
         });
+    }
+
+    private _emitModelValueChange() {
+        if (this.options) {
+            let selectedValues = this.selectionModel.selected;
+            if (this.thyUniqueKey) {
+                selectedValues = selectedValues.map((selectedValue) => {
+                    const selectedOption = this.options.find((option) => {
+                        return option.thyValue[this.thyUniqueKey] === selectedValue;
+                    });
+                    return selectedOption.thyValue;
+                });
+            }
+            let changeValue = selectedValues;
+            if (!this.multiple && selectedValues && selectedValues.length > 0) {
+                changeValue = selectedValues[0];
+            }
+            this._onChange(changeValue);
+        }
     }
 
     private _toggleFocusedOption(event: KeyboardEvent): void {
@@ -112,7 +136,7 @@ export class ThySelectionListComponent implements OnInit, OnDestroy, AfterConten
     }
 
     private _initializeFocusKeyManager() {
-        this._keyManager = new FocusKeyManager<ThyListOptionComponent>(this.options)
+        this._keyManager = new ActiveDescendantKeyManager<ThyListOptionComponent>(this.options)
             .withWrap()
             // .withTypeAhead()
             // Allow disabled items to be focusable. For accessibility reasons, there must be a way for
@@ -120,56 +144,40 @@ export class ThySelectionListComponent implements OnInit, OnDestroy, AfterConten
             .skipPredicate(() => false);
     }
 
-    private _subscribeSelectionModelChanges() {
-        // Sync external changes to the model back to the options.
-        this._selectionChangesUnsubscribe$ = this.selectionModel.onChange.subscribe(event => {
-            if (event.added) {
-                for (const item of event.added) {
-                    item.selected = true;
-                }
-            }
-
-            if (event.removed) {
-                for (const item of event.removed) {
-                    item.selected = false;
-                }
-            }
-        });
-    }
-
     private _instanceSelectionModel() {
-        this.selectionModel = new SelectionModel<ThyListOptionComponent>(this.multiple);
+        this.selectionModel = new SelectionModel<any>(this.multiple);
     }
 
-    private _getSelectorElement(element: HTMLElement | ElementRef | string) {
-        if (!element) {
-            return this.elementRef.nativeElement;
-        } else if (element === 'body') {
-            return this.document;
-        } else if (helpers.isString(element)) {
-            return this.document.querySelector(element as string);
-        } else if (element instanceof ElementRef) {
-            return element.nativeElement;
+    private _getElementBySelector(element: HTMLElement | ElementRef | string): HTMLElement {
+        return dom.getHTMLElementBySelector(element, this.elementRef);
+    }
+
+    private _compareValue(value1: any, value2: any) {
+        if (this.thyCompareWith) {
+            const compareFn = this.thyCompareWith as (o1: any, o2: any) => boolean;
+            return compareFn(value1, value2);
+        } else if (this.thyUniqueKey) {
+            return value1 && value1[this.thyUniqueKey] === value2 && value2[this.thyUniqueKey];
         } else {
-            return element;
+            return value1 === value2;
         }
     }
 
-    private _setOptionsFromValues(values: any[]) {
-        // this.options.forEach(option => option.setSelected(false));
+    private _getOptionSelectionValue(option: ThyListOptionComponent) {
+        if (option.thyValue) {
+            return this.thyUniqueKey ? option.thyValue[this.thyUniqueKey] : option.thyValue;
+        } else {
+            return option;
+        }
+    }
+
+    private _setSelectionByValues(values: any[]) {
         this.selectionModel.clear();
         values.forEach(value => {
-            const correspondingOption = this.options.find(option => {
-                // Skip options that are already in the model. This allows us to handle cases
-                // where the same primitive value is selected multiple times.
-                if (option.selected) {
-                    return false;
-                }
-
-                return this.thyCompareWith ? this.thyCompareWith(option.thyValue, value) : option.thyValue === value;
-            });
-            if (correspondingOption) {
-                this.selectionModel.select(correspondingOption);
+            if (this.thyUniqueKey) {
+                this.selectionModel.select(value[this.thyUniqueKey]);
+            } else {
+                this.selectionModel.select(value);
             }
         });
     }
@@ -180,28 +188,29 @@ export class ThySelectionListComponent implements OnInit, OnDestroy, AfterConten
         let hasChanged = false;
 
         this.options.forEach(option => {
-            const fromIsSelected = this.selectionModel.isSelected(option);
+            const fromIsSelected = this.selectionModel.isSelected(option.thyValue);
             if (fromIsSelected !== toIsSelected) {
                 hasChanged = true;
-                this.selectionModel.toggle(option);
+                this.selectionModel.toggle(option.thyValue);
             }
         });
 
         if (hasChanged) {
-            this._reportModelValueChange();
+            this._emitModelValueChange();
         }
     }
 
-    private _reportModelValueChange() {
-        if (this.options) {
-            const selectedValues = this.selectionModel.selected.map((option) => {
-                return option.thyValue;
-            });
-            let changeValue = selectedValues;
-            if (!this.multiple && selectedValues && selectedValues.length > 0) {
-                changeValue = selectedValues[0];
-            }
-            this._onChange(changeValue);
+    private _getOptionByValue(value: any) {
+        return this.options.find(option => {
+            return this._compareValue(option.thyValue, value);
+        });
+    }
+
+    private _getActiveOption() {
+        if (this._keyManager.activeItem) {
+            return this._getOptionByValue(this._keyManager.activeItem.thyValue);
+        } else {
+            return null;
         }
     }
 
@@ -214,12 +223,11 @@ export class ThySelectionListComponent implements OnInit, OnDestroy, AfterConten
     }
 
     ngOnInit() {
-        const bindKeyEventElement = this._getSelectorElement(this.thyBindKeyEventContainer);
+        const bindKeyEventElement = this._getElementBySelector(this.thyBindKeyEventContainer);
         this.ngZone.runOutsideAngular(() => {
             this._bindKeyEventUnsubscribe = this.renderer.listen(bindKeyEventElement, 'keydown', this.onKeydown.bind(this));
         });
         this._instanceSelectionModel();
-        this._subscribeSelectionModelChanges();
     }
 
     writeValue(value: any[] | any): void {
@@ -231,9 +239,9 @@ export class ThySelectionListComponent implements OnInit, OnDestroy, AfterConten
                 throw new Error(`single selection ngModel not be array.`);
             }
         }
-        const values = helpers.isArray(value) ? value : (value ? [value] : null);
+        const values = helpers.isArray(value) ? value : (value ? [value] : []);
         if (this.options) {
-            this._setOptionsFromValues(values || []);
+            this._setSelectionByValues(values);
         } else {
             this._tempValues = values;
         }
@@ -273,7 +281,6 @@ export class ThySelectionListComponent implements OnInit, OnDestroy, AfterConten
             default:
                 manager.onKeydown(event);
         }
-
         if ((keyCode === keycodes.UP_ARROW || keyCode === keycodes.DOWN_ARROW) && event.shiftKey &&
             manager.activeItemIndex !== previousFocusIndex) {
             this._toggleFocusedOption(event);
@@ -282,16 +289,37 @@ export class ThySelectionListComponent implements OnInit, OnDestroy, AfterConten
 
     toggleOption(option: ThyListOptionComponent, event?: Event) {
         if (option && !option.disabled) {
-            this.selectionModel.toggle(option);
+            this.selectionModel.toggle(this._getOptionSelectionValue(option));
             // Emit a change event because the focused option changed its state through user
             // interaction.
-            this._reportModelValueChange();
+            this._emitModelValueChange();
             this._emitChangeEvent(option, event);
         }
     }
 
-    setFocusedOption(option: ThyListOptionComponent) {
+    setActiveOption(option: ThyListOptionComponent) {
         this._keyManager.updateActiveItem(option); // .updateActiveItemIndex(this._getOptionIndex(option));
+    }
+
+    scrollIntoView(option: ThyListOptionComponent) {
+        const scrollContainerElement = dom.getHTMLElementBySelector(this.thyScrollContainer, this.elementRef);
+        ScrollToService.scrollToElement(option.element.nativeElement, scrollContainerElement);
+    }
+
+    isSelected(option: ThyListOptionComponent) {
+        return this.selectionModel.isSelected(this._getOptionSelectionValue(option));
+    }
+
+    clearActiveItem() {
+        if (this._keyManager.activeItem) {
+            this._keyManager.setActiveItem(-1);
+        }
+    }
+
+    determineClearActiveItem() {
+        if (!this._getActiveOption()) {
+            this.clearActiveItem();
+        }
     }
 
     /** Selects all of the options. */
@@ -306,9 +334,8 @@ export class ThySelectionListComponent implements OnInit, OnDestroy, AfterConten
 
     ngAfterContentInit(): void {
         this._initializeFocusKeyManager();
-        // this._subscribeSelectionModelChanges();
         if (this._tempValues) {
-            this._setOptionsFromValues(this._tempValues);
+            this._setSelectionByValues(this._tempValues);
             this._tempValues = null;
         }
     }